Extract AbstractBletia from Bletia
